chore: move shared replace feature in dedicated sub-directory
