+++ lib/camunda-cloud/Modeler.js
@@ -13,7 +13,7 @@ import {
 } from 'bpmn-js-properties-panel';
 
 import replaceModule from './features/replace';
-import sharedModules from '../shared';
+import sharedReplaceModule from '../shared/features/replace';
 
 import { commonModdleExtensions, commonModules } from './util/commonModules';
 
@@ -48,7 +48,7 @@ Modeler.prototype._camundaCloudModules = [
   zeebePropertiesProviderModule,
   cloudElementTemplatesPropertiesProvider,
   replaceModule,
-  sharedModules
+  sharedReplaceModule
 ];
 
 Modeler.prototype._modules = [].concat(
