+++ lib/camunda-platform/Modeler.js
@@ -11,7 +11,7 @@ import {
 
 import { commonModdleExtensions } from './util/commonModules';
 
-import sharedModules from '../shared';
+import sharedReplaceModule from '../shared/features/replace';
 
 /**
  *
@@ -36,7 +36,7 @@ Modeler.prototype._camundaPlatformModules = [
   behaviorsModule,
   CamundaPlatformPropertiesProviderModule,
   ElementTemplatesPropertiesProviderModule,
-  sharedModules
+  sharedReplaceModule
 ];
 
 Modeler.prototype._modules = [].concat(
