+++ lib/shared/features/replace/util/ReplaceOptionsUtil.js
